+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Coinbase Node.js SDK Changelog
 
-## Unreleased
+## Unreleased 
 
+## [0.5.0] - 2024-09-11
+
+- Add Arbitrum-Mainnet support for Native transfers.
 - Add `Coinbase.configure` method to allow for configuration of the SDK and marked constructor as deprecated.
 - Return correlation ID from APIError response
 - Add optional fields to `CreateContractInvocationOptions` to set amount for payable contract method invocations